prov/gni: Abstracted GNIX MR Cache and added
callbacks

The GNIX cache has been abstracted so that it can be
used for other types of registrations such as XPMEM.
Callbacks have been added to support device specific
functionality for registration, deregistration and
object destruction.

(cherry picked from commit ofi-cray/libfabric-cray@589afd199065a5746bf5d07fc2402354aa4bcce5)
upstream merge of ofi-cray/libfabric-cray#664
